refactor(tests): extract utils from NoteTypeTest
