+++ libanki/testutils/src/main/java/com/ichi2/anki/libanki/testutils/ext/Collection.kt
@@ -16,10 +16,48 @@
 
 package com.ichi2.anki.libanki.testutils.ext
 
+import anki.notetypes.StockNotetype
 import com.ichi2.anki.libanki.Collection
 import com.ichi2.anki.libanki.Note
+import com.ichi2.anki.libanki.NotetypeJson
+import com.ichi2.anki.libanki.addNotetypeLegacy
+import com.ichi2.anki.libanki.backend.BackendUtils
+import com.ichi2.anki.libanki.getStockNotetype
+
+const val BASIC_NOTE_TYPE_NAME = "Basic"
 
 fun Collection.addNote(note: Note): Int {
     addNote(note, note.notetype.did)
     return note.numberOfCards(this)
+}
+
+/**
+ * Creates a basic model.
+ *
+ * Note: changes to this model will propagate to [createBasicTypingNoteType] as that model is built on
+ * top of the model returned by this function.
+ *
+ * @param name name of the new model
+ * @return the new model
+ */
+fun Collection.createBasicNoteType(name: String = BASIC_NOTE_TYPE_NAME): NotetypeJson {
+    val noteType =
+        getStockNotetype(StockNotetype.Kind.KIND_BASIC).apply { this.name = name }
+    addNotetypeLegacy(BackendUtils.toJsonBytes(noteType))
+    return notetypes.byName(name)!!
+}
+
+/**
+ * Creates a basic typing model.
+ *
+ * @see createBasicNoteType
+ */
+fun Collection.createBasicTypingNoteType(name: String): NotetypeJson {
+    val noteType = createBasicNoteType(name)
+    noteType.templates[0].apply {
+        qfmt = "{{Front}}\n\n{{type:Back}}"
+        afmt = "{{Front}}\n\n<hr id=answer>\n\n{{type:Back}}"
+    }
+    notetypes.save(noteType)
+    return noteType
 }
